requirements.txt
colorama==0.4.4
discord.py==1.7.3
discord==1.7.3
python-dotenv==0.19.0